requirements.txt
beautifulsoup4==4.6.1
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
et-xmlfile==1.0.1
get==1.0.3
idna==2.7
jdcal==1.4
kiwisolver==1.0.1
matplotlib==2.2.0
numpy==1.14.2
openpyxl==2.5.4
pandas==0.22.0
post==1.0.2
public==1.0.3
pyparsing==2.2.0
python-dateutil==2.7.0
pytz==2018.3
PyYAML==3.13
query-string==1.0.2
requests==2.19.1
scipy==1.1.0
six==1.11.0
urllib3==1.23
xlrd==1.1.0
